requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
google-auth==2.48.0
google-genai==1.63.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.6.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
pillow==12.1.1
pytesseract==0.3.13
pdf2image==1.17.0
python-docx==0.8.11
PyMuPDF==1.23.0
